requirements.txt
flask
flask-sqlalchemy
flask-cors
scikit-learn
tensorflow
keras
psycopg2-binary
flask-migrate
python-dotenv
gunicorn
numpy
h5py
